refactor: regroup service deps - organize service dependeny abstractions into files - align related tests
