+++ packages/cardano-services/src/Program/index.ts
@@ -4,3 +4,4 @@ export * from './loadHttpServer';
 export * from './ProgramOptionDescriptions';
 export * from './ServiceNames';
 export * from './utils';
+export * from './services';
